refactored to split up code
