+++ react-app/src/components/NavBar/index.js
@@ -6,61 +6,10 @@ import SearchBar from "../SearchBar";
 import { useSelector } from "react-redux";
 import { useState } from "react";
 import Modal from "../Modals";
+import UserControls from "./userControls";
 const NavBar = () => {
-  const session = useSelector((state) => state.session);
-  const [toggleUserInfo, setToggleUserInfo] = useState(true);
   const [modal, setModal] = useState(false);
-  function userControls() {
-    if (session?.user) {
-      return (
-        <>
-          <Switch>
-            <Route path="/favorites">
-              <Link
-                to="/"
-                className="material-icons"
-                style={{
-                  color: "#8a2be2",
-                  width: "24px",
-                }}
-              >
-                favorite
-              </Link>
-            </Route>
-            <Route path="/">
-              <Link
-                to="/favorites"
-                className="material-icons"
-                style={{
-                  color: "white",
-                  width: "24px",
-                }}
-              >
-                favorite_outlined
-              </Link>
-            </Route>
-          </Switch>
 
-          <button
-            className={style.user}
-            onClick={() => setToggleUserInfo(!toggleUserInfo)}
-          >
-            <span className="material-icons">account_circle</span>
-            <span className="material-icons">arrow_drop_down</span>
-          </button>
-          <button className={style.cart}>
-            <span className="material-icons">shopping_cart</span>
-          </button>
-        </>
-      );
-    } else {
-      return (
-        <button className={style.signIn} onClick={() => setModal(true)}>
-          Sign in
-        </button>
-      );
-    }
-  }
   return (
     <nav>
       <div className={style.content}>
@@ -72,7 +21,9 @@ const NavBar = () => {
           <img src={logo_words} alt="" />
         </NavLink>
         <SearchBar />
-        <div className={style.userControls}>{userControls()}</div>
+        <div className={style.userControls}>
+          <UserControls setModal={setModal} />
+        </div>
       </div>
       {modal && <Modal setModal={setModal} />}
     </nav>
