+++ react-app/src/components/NavBar/navbar.module.css
@@ -31,6 +31,7 @@ nav {
   outline: inherit;
   height: 24px;
 }
+
 .userControls {
   display: flex;
   align-items: center;
@@ -38,13 +39,41 @@ nav {
   margin-left: 1rem;
 }
 .user {
+  position: relative;
   height: 24px;
   display: flex;
   align-items: center;
 }
-.user span:last-child {
+.user > img {
+  height: 100%;
+  border-radius: 50%;
+}
+.user span:last-of-type {
   font-size: 18px;
 }
 .userControls a {
   height: 24px;
+}
+.userDropdown {
+  position: absolute;
+  top: 40px;
+  right: -10px;
+  width: 230px;
+  height: 280px;
+  background-color: rgb(29, 29, 29);
+  box-shadow: 0 0 4px black;
+  border-radius: 1rem;
+}
+
+.userDropdown::after {
+  content: "";
+  position: absolute;
+  width: 0;
+  height: 0;
+  top: -8px;
+  right: 20px;
+  border-left: 8px solid transparent;
+  border-right: 8px solid transparent;
+
+  border-bottom: 8px solid rgb(29, 29, 29);
 }
