+++ react-app/src/components/SearchBar/searchbar.module.css
@@ -1,5 +1,5 @@
 .container {
-  width: 45%;
+  width: 65%;
   position: relative;
 }
 .search {
